Extract score check into its own file
