+++ src/domain-logic/score-update.js
@@ -1,21 +1,4 @@
-export const WIN = 'WIN'
-export const ADV = 'ADV'
-export const DEUCE_SCORE = 40
-
-const isDeuce = currentScore => 
-  Object
-    .values(currentScore)
-    .every(score => score === DEUCE_SCORE)
-
-const isAdv = currentScore => 
-  Object
-    .values(currentScore)
-    .some(score => score === ADV)
-
-const isGameEnded = currentScore => 
-  Object
-    .values(currentScore)
-    .some(score => score === WIN)
+import { WIN, ADV, DEUCE_SCORE, isDeuce, isAdv, isGameEnded } from './score-check'
 
 const normalUpdate = (currentScore, scorer) => {
   const nextScoreOf = {
